Minor refactoring into separate files, add some more comments, fix permission issue, fix bug in AllFiles returning directories.
